rename to props
