+++ src/components/common.test.ts
@@ -11,8 +11,8 @@ import {
   PositionProps,
   ShadowProps,
 } from "./common";
-import { VisibilityProps } from "@/prop/shared/visibility.props";
-import { DimensionProps } from "@/prop/shared/dimension.props";
+import { VisibilityProps } from "@/props/shared/visibility.props";
+import { DimensionProps } from "@/props/shared/dimension.props";
 
 describe("Flex", () => {
   afterEach(() => {
