+++ src/components/layout/ScrollArea/ScrollArea.tsx
@@ -6,7 +6,7 @@ import {
   WrapperProps,
 } from "../../common";
 import styles from "./ScrollArea.module.css";
-import { DimensionProps } from "@/prop/shared/dimension.props";
+import { DimensionProps } from "@/props/shared/dimension.props";
 
 type ScrollBarDirection = "horizontal" | "vertical" | "both";
