+++ src/props/main.props.ts
