+++ src/props/shared/dimension.props.ts
