+++ src/props/shared/visibility.props.ts
